requirement.txt
django==1.9.7
pillow==3.2.0
django-updown
